+++ NEWS.md
@@ -1,24 +1,32 @@
 # daisugi (development)
-* add support for **ebm**
+
+# daisugi 0.0.3
+* added support for **EBMs**
+    - fit with `grow_explainable_trees`
+    - predict with `harvest_explainable_trees`
+    - a 'glassbox' novel gam boosting machine
+* added support for **NGBoost**
     - fit with `grow_explainable_trees`
     - predict with `harvest_explainable_trees`
     - a 'glassbox' novel gam boosting machine
 
+
+
 # daisugi 0.0.2
 
-* added support for **wildwood**
+* added support for **WildWood**
     - fit with `grow_wild_trees`
     - predict with `harvest_wild_trees`
     - a *newer and advanced* random forest algorithm 
 
-* added support for **perpetual**
+* added support for **Perpetual**
     - fit with `grow_perpetual_trees`
     - predict with `harvest_perpetual_trees`
     - a budget-based autoML-type tree machine
 
 # daisugi 0.0.1
 
-* added support for **yggdrasil decision forest**
+* added support for **Yggdrasil Decision Forest**
     - fit with `grow_yggdrasil_trees`
     - predict with `harvest_yggdrasil_trees`
     - YDF Gradient Boosted Trees (Google)